// backend/tournament-service/src/types/enums.ts
export enum GameType {
  CHESS = 'chess',
  POKER = 'poker',
  BACKGAMMON = 'backgammon',
  GO = 'go',
}

export enum TournamentType {
  DAILY = 'daily',
  WEEKLY = 'weekly',
  MONTHLY = 'monthly',
}

export enum TournamentStatus {
  OPEN = 'open',
  FULL = 'full',
  IN_PROGRESS = 'in_progress',
  COMPLETED = 'completed',
}